refactor: 按功能拆包并补上可离线的接口测试
将 mtop/IM/路由从根目录 spider.py 拆到 xianyu 包，方便继续加登录与自动回复。
补充 FastAPI TestClient 测试；真实搜索用 RUN_LIVE=1 可选执行。
