+++ xianyu/__init__.py
@@ -1,0 +1,1 @@
+"""闲鱼 HTTP 客户端：搜索、登录、IM 自动回复。"""
