+++ xianyu/config.py
@@ -1,0 +1,13 @@
+import os
+from pathlib import Path
+
+ROOT_DIR = Path(__file__).resolve().parent.parent
+DATA_DIR = ROOT_DIR / "data"
+DATA_DIR.mkdir(parents=True, exist_ok=True)
+
+DATABASE_URL = os.environ.get("DATABASE_URL") or f"sqlite://{DATA_DIR / 'xianyu.sqlite3'}"
+SESSION_PATH = DATA_DIR / "session.json"
+UA = (
+    "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 "
+    "(KHTML, like Gecko) Chrome/120.0.0.0 Safari/537.36"
+)
