+++ xianyu/session.py
@@ -3,10 +3,9 @@
 from __future__ import annotations
 
 import json
-from pathlib import Path
 from typing import Optional
 
-SESSION_PATH = Path("data/session.json")
+from xianyu.config import SESSION_PATH
 
 
 def load_session() -> Optional[dict]:
